refactor: Combine individual files for functions and tests into single files
