--- src/functions/get-random-user.js
@@ -1,8 +1,0 @@
-function getRandomUser(users) {
-    const randomIndex = Math.floor(Math.random() * users.length);
-    return users[randomIndex];
-}
-
-export {
-    getRandomUser,
-}
